move styles to SedaMetricSlider
