+++ src/modules/explorer/filters/components/SedaMetricSlider.js
@@ -10,6 +10,17 @@ import { DebouncedSlider as Slider } from '../../../../shared'
 import { getPrefixLang } from '../../app/selectors/lang'
 import { getFormatterForVarName } from '../../app/selectors'
 import useActiveFilters from '../hooks/useActiveFilters'
+import { Tooltip, withStyles } from '@material-ui/core'
+
+const ValueLabelComponent = (props) => {
+  const { children, open, value } = props;
+
+  return (
+    <Tooltip open={open} enterTouchDelay={0} placement="top" arrow title={value}>
+      {children}
+    </Tooltip>
+  );
+}
 
 const getDefaultValue = (region, metric) => {
   const vals = DEFAULT_RANGES[region]
@@ -17,7 +28,7 @@ const getDefaultValue = (region, metric) => {
   return vals[metric]
 }
 
-const SedaMetricSlider = ({ metricId, ...props }) => {
+const SedaMetricSlider = ({ metricId, classes, isActive, ...props }) => {
   // grab filters array
   const filters = useActiveFilters()
   // function to remove a single filter
@@ -66,7 +77,7 @@ const SedaMetricSlider = ({ metricId, ...props }) => {
       titleProps={{ id: metricId + '-slider' }}
       {...props}>
       <Slider
-        metric={metricId}
+        ValueLabelComponent={ValueLabelComponent}
         marks={marks}
         valueLabelDisplay="auto"
         value={value}
@@ -78,6 +89,7 @@ const SedaMetricSlider = ({ metricId, ...props }) => {
         aria-labelledby={metricId + '-slider'}
         getAriaValueText={formatter}
         valueLabelFormat={formatter}
+        classes={classes}
       />
     </PanelListItem>
   )
@@ -85,4 +97,49 @@ const SedaMetricSlider = ({ metricId, ...props }) => {
 
 SedaMetricSlider.propTypes = {}
 
-export default SedaMetricSlider
+export default withStyles(theme => ({
+  root: {
+    padding: '10px 0',
+    position: 'relative'
+  },
+  thumb: {
+    height: 12,
+    width: 6,
+    marginTop: -2,
+    marginLeft: -3,
+    border: '0.5px solid #757575',
+    background: 'white',
+    borderRadius: 2,
+    '&:focus, &:hover, &$active': {
+      height: 14,
+      marginTop: -3,
+      borderColor: 'currentColor',
+      boxShadow: 'inherit'
+    },
+  },
+  track: {
+    height: 8,
+    background: props => !props.isActive ? 'currentColor' : 'fixed linear-gradient(90deg, #1C4E81 100px, #3A7BBB 139px, #7CB4DC 178px, #C0E1EC 217px, #F5F6F7 256px, #D6EECA 295px, #9FDCA3 334px, #32A877 373px, #1B704D 412px)',
+    [theme.breakpoints.down('sm')]: {
+      background: props => !props.isActive ? 'currentColor' : 'fixed linear-gradient(90deg, #1C4E81 20px, #3A7BBB 59px, #7CB4DC 98px, #C0E1EC 137px, #F5F6F7 176px, #D6EECA 215px, #9FDCA3 254px, #32A877 293px, #1B704D 332px)',
+    }
+  },
+  rail: {
+    height: 8,
+    background: '#E0E0E0',
+    opacity: 1
+  },
+  mark: {
+    height: 5,
+    width: 1,
+    backgroundColor: '#E0E0E0',
+    marginTop: 13
+  },
+  markLabel: {
+    color: '#5D5D5D',
+    marginTop: 2,
+    [theme.breakpoints.down('sm')]: {
+      top: 26
+    }
+  },
+}))(SedaMetricSlider)
