+++ src/shared/components/Inputs/Slider.js
@@ -1,72 +1,50 @@
 import React, { useState, useRef } from 'react'
 import {
   withStyles,
-  Slider as MuiSlider,
-  Tooltip
+  Slider as MuiSlider
 } from '@material-ui/core'
 import * as _debounce from 'lodash.debounce'
 import useDidUpdateEffect from '../../hooks/useDidUpdateEffect'
 import shallow from 'zustand/shallow'
 import usePrevious from '../../hooks/usePrevious'
 
-const StyledSlider = withStyles(theme => ({
+const StyledSlider = withStyles({
   root: {
     height: 2,
-    padding: '10px 0',
-    position: 'relative'
+    padding: '16px 0'
   },
   thumb: {
-    height: 12,
-    width: 6,
-    marginTop: -2,
-    marginLeft: -3,
-    border: '0.5px solid #757575',
-    background: 'white',
-    borderRadius: 2,
-    '&:focus, &:hover, &$active': {
-      height: 14,
-      marginTop: -3,
-      borderColor: 'currentColor',
-      boxShadow: 'inherit'
-    },
+    height: 16,
+    width: 16,
+    marginTop: -8,
+    marginLeft: -8
   },
   active: {},
-  track: {
-    height: 8,
-    background: (props) => props.metric === 'ses' ? 'currentColor' : 'fixed linear-gradient(90deg, #1C4E81 100px, #3A7BBB 139px, #7CB4DC 178px, #C0E1EC 217px, #F5F6F7 256px, #D6EECA 295px, #9FDCA3 334px, #32A877 373px, #1B704D 412px)',
-    [theme.breakpoints.down('sm')]: {
-      background: (props) => props.metric === 'ses' ? 'currentColor' : 'fixed linear-gradient(90deg, #1C4E81 20px, #3A7BBB 59px, #7CB4DC 98px, #C0E1EC 137px, #F5F6F7 176px, #D6EECA 215px, #9FDCA3 254px, #32A877 293px, #1B704D 332px)',
+  valueLabel: {
+    left: 'calc(-50%)',
+    top: 22,
+    '& *': {
+      background: 'transparent',
+      color: '#000'
     }
+  },
+  track: {
+    height: 2
   },
   rail: {
-    height: 8,
-    background: '#E0E0E0',
-    opacity: 1
+    height: 2,
+    opacity: 0.5
   },
   mark: {
-    height: 5,
+    height: 8,
     width: 1,
-    backgroundColor: '#E0E0E0',
-    marginTop: 13
+    marginTop: -3
   },
-  markLabel: {
-    color: '#5D5D5D',
-    marginTop: 2,
-    [theme.breakpoints.down('sm')]: {
-      top: 26
-    }
-  },
-}))(MuiSlider)
-
-const ValueLabelComponent = (props) => {
-  const { children, open, value } = props;
-
-  return (
-    <Tooltip open={open} enterTouchDelay={0} placement="top" arrow title={value}>
-      {children}
-    </Tooltip>
-  );
-}
+  markActive: {
+    opacity: 1,
+    backgroundColor: 'currentColor'
+  }
+})(MuiSlider)
 
 /**
  * Gets the value for the slider, or provides the default value if none is given
@@ -129,7 +107,6 @@ export function DebouncedSlider({
   }
   return (
     <StyledSlider
-      ValueLabelComponent={ValueLabelComponent}
       value={value}
       onChange={handleChange}
       {...props}
